Title time invariant two-sex multi-state kin projections

Title time invariant two-sex multi-state kin projections

all_kin_dy(
  Uf,
  Um,
  Ff,
  Fm,
  alpha,
  na,
  ns,
  Parity,
  sex_Focal,
  Initial_stage_Focal
)

Arguments

Uf

matrix (block structured). transfers female individuals across stages and advances their age (conditional on survial)

Um

matrix (block structured). transfers male individuals across stages and advances their age (conditional on survial)

Ff

matrix (block structured). accounts for female reproduction, and assigns newborns into given age*stage

Fm

matrix (block structured). accounts for male reproduction; assigns newborns into age-class, and stage

alpha

scalar. birth ratio (male:female)

na

scalar. number of ages.

ns

scalar. number of stages.

Parity

logical. If true then we omit mothers of parity 0, and re-scale the mother's age*stage of parenting

sex_Focal

logical. Female or Male

Initial_stage_Focal

numeric. Any natural number 1,2,3,4,...

Value

a list of matrices. Each list entry represents a particular kin. Each kin is chacacterised by a matrix of dimension: nrow = 2* na * ns (2-sex age-stage structured) and ncol = na (Focal's age) yielding the age*stage distribution of kin for each age of Focal