Title time-variant two-sex multi-state kin projections

Title time-variant two-sex multi-state kin projections

Usage

all_kin_dy_TV(
  Uf,
  Um,
  Ff,
  Fm,
  alpha,
  na,
  ns,
  Parity,
  sex_Focal,
  Initial_stage_Focal,
  previous_kin_Focal,
  prev_kin_children,
  prev_kin_grandchildren,
  prev_kin_greatgrandchildren,
  prev_kin_parents,
  prev_kin_grand_parents,
  prev_kin_great_grand_parents,
  prev_kin_older_sibs,
  prev_kin_younger_sibs,
  prev_kin_older_niece_nephew,
  prev_kin_younger_niece_nephew,
  prev_kin_older_aunts_uncles,
  prev_kin_younger_aunts_uncles,
  prev_kin_older_cousins,
  prev_kin_younger_cousins,
  previous_population_age_stage_structure
)

Arguments

Uf

matrix (block structured). transfers female individuals across stages and advances their age (conditional on survial)

Um

matrix (block structured). transfers male individuals across stages and advances their age (conditional on survial)

Ff

matrix (block structured). accounts for female reproduction, and assigns newborns into given age*stage

Fm

matrix (block structured). accounts for male reproduction; assigns newborns into age-class, and stage

alpha

scalar. birth ratio (male:female)

na

scalar. number of ages.

ns

scalar. number of stages.

Parity

logical. If true then we omit mothers of parity 0, and re-scale the mother's age*stage of parenting

sex_Focal

logical. Female or Male

Initial_stage_Focal

numeric. Any natural number 1,2,3,4,...

previous_kin_Focal

matrix. last years kinship output.

prev_kin_children

matrix. last years kinship output.

prev_kin_grandchildren

matrix. last years kinship output.

prev_kin_greatgrandchildren

matrix. last years kinship output.

prev_kin_parents

matrix. last years kinship output.

prev_kin_grand_parents

matrix. last years kinship output.

prev_kin_older_sibs

matrix. last years kinship output.

prev_kin_younger_sibs

matrix. last years kinship output.

prev_kin_older_niece_nephew

matrix. last years kinship output.

prev_kin_younger_niece_nephew

matrix. last years kinship output.

prev_kin_older_aunts_uncles

matrix. last years kinship output.

prev_kin_younger_aunts_uncles

matrix. last years kinship output.

prev_kin_older_cousins

matrix. last years kinship output.

prev_kin_younger_cousins

matrix. last years kinship output.

previous_population_age_stage_structure

vector. The transient "population structure" (age*stage distributed)

Value

a list of matrices. Each list entry represents a particular kin. Each kin is chacacterised by a matrix of dimension: nrow = 2* na * ns (2-sex age-stage structured) and ncol = na (Focal's age) yielding the age*stage distribution of kin for each age of Focal