Estimate kin counts in a time invariant framework for two-sex model.

Two-sex matrix framework for kin count estimates.This produces kin counts grouped by kin, age and sex of each relatives at each Focal´s age. For example, male cousins from aunts and uncles from different sibling's parents are grouped in one male count of cousins.

Usage

kin_time_invariant_2sex(
  pf = NULL,
  pm = NULL,
  ff = NULL,
  fm = NULL,
  sex_focal = "f",
  birth_female = 1/2.04,
  pif = NULL,
  pim = NULL,
  output_kin = NULL,
  list_output = FALSE
)

Arguments

pf

numeric. A vector of survival probabilities for females with same length as ages.

pm

numeric. A vector of survival probabilities for males with same length as ages.

ff

numeric. A vector of age-specific fertility rates for females with same length as ages.

fm

numeric. A vector of age-specific fertility rates for males with same length as ages.

sex_focal

character. "f" for female or "m" for male.

birth_female

numeric. Female portion at birth.

pif

numeric. For using some specific non-stable age distribution of childbearing for mothers (same length as ages). Default NULL.

pim

numeric. For using some specific non-stable age distribution of childbearing for fathers (same length as ages). Default NULL.

output_kin

character. kin to return, considering matrilineal names. For example "m" for parents, "d" for children, etc. See the vignette for all kin types.

list_output

logical. Results as a list with output_kin elements, with focal´s age in columns and kin ages in rows (2 * ages, last chunk of ages for death experience). Default FALSE

Value

A data frame with focal´s age, related ages and type of kin (for example d is children, oa is older aunts/uncles, etc.), sex, alive and death. If list_output = TRUE then this is a list.

Details

See Caswell (2022) for details on formulas.