APC combination to return

define APC combination to return in kin and kin2sex.

Usage

output_period_cohort_combination(
  output_cohort = NULL,
  output_period = NULL,
  age = NULL,
  years_data = NULL
)

Arguments

output_cohort

integer. A vector with selected calendar years.

output_period

integer. A vector with selected cohort years.

age

integer. A vector with ages from the kinship network to be filtered.

years_data

integer. A vector with years from the time-varying kinship network to be filtered.

Value

data.frame with years and ages to filter in kin and kin_2sex functions.

Details

Because returning all period and cohort data from a huge time-series would be hard memory consuming, this function is an auxiliary one to deal with selection from inputs output_cohort and output_period.