Mixing distributions for the time-invariant multi-state 2-sex model: Non-parity case

Mixing distributions for the time-invariant multi-state 2-sex model: Non-parity case

Usage

pi_mix(Uf, Um, Ff, Fm, alpha, na, ns)

Arguments

Uf

matrix. Block-structured matrix which transfers females over stage and advances their age

Um

matrix. Block-structured matrix which transfers males over stage and advances their age

Ff

matrix. Block-structured matrix which counts reproduction by females and assigns newborns an age and stage

Fm

matrix. Block-structured matrix which counts reproduction by males and assigns newborns an age and stage

alpha

scalar. Birth ratio male:female

na

scalar. Number of age-classes

ns

scalar. Number of stages

Value

list (of vectors). list[1] = full agestagesex distribution list[2] = female agestage distribution normalised list[3] = male agestage distribution normalised list[4] = female marginal age distribution normalised list[5] = male marginal age distribution normalised