Mixing distributions for the time-variant multi-state 2-sex model: Non-parity case

Mixing distributions for the time-variant multi-state 2-sex model: Non-parity case

Usage

pi_mix_TV(Ff, Fm, alpha, na, ns, previous_age_stage_dist)

Arguments

Ff

matrix. Block-structured matrix which counts reproduction by females and assigns newborns an age and stage

Fm

matrix. Block-structured matrix which counts reproduction by males and assigns newborns an age and stage

alpha

scalar. Birth ratio male:female

na

scalar. Number of age-classes

ns

scalar. Number of stages

previous_age_stage_dist

vector. Last years population structure (agestagesex full distribution)

Value

list (of vectors). list[1] = full agestagesex distribution list[2] = female agestage distribution normalised list[3] = male agestage distribution normalised list[4] = female marginal age distribution normalised list[5] = male marginal age distribution normalised