plot a Kin diagram (network)

Draws a Keyfitz-style kinship diagram given a kinship object created by the kin function. Displays expected kin counts for a Focal aged 'a'.

Usage

plot_diagram(kin_total, rounding = 3)

Arguments

kin_total

data.frame. values in column kin define the relative type - see demokin_codes(). Values in column count are the expected number of relatives.

rounding

numeric. Number of decimals to show in diagram.

Value

A Keyfitz-style kinship plot.