one time projection kin

one time projection kin. internal function.

Usage

timevarying_kin_2sex(Ut, Ft, Ft_star, pit, sex_focal, ages, pkin)

Arguments

Ut

numeric. A matrix of survival probabilities (or ratios).

Ft

numeric. A matrix of age-specific fertility rates.

Ft_star

numeric. Ft but for female fertility.

pit

numeric. A matrix with distribution of childbearing.

sex_focal

character. "f" for female or "m" for male.

ages

numeric.

pkin

numeric. A list with kin count distribution in previous year.

Value

A list of 14 types of kin matrices (kin age by Focal age, blocked for two sex) projected one time interval.